requirements/prod.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.

Django==1.7.7
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.17
boto==2.9.0
cef==0.5
commonware==0.4.0
curling==0.3.9
dj-database-url==0.2.1
django-aesfield==0.1.2
django-extensions==0.9
django-filter==0.6
django-multidb-router==0.5
django-nose==1.3
django-paranoia==0.1.7
django-raven-metlog==0.1
django-sha2==0.4
django-statsd-mozilla==0.3.8.3
djangorestframework==2.3.9
gunicorn==0.17.2
httplib2==0.7.6
lxml==2.2.6
m2secret==0.1.1
metlog-cef==0.2
metlog-py==0.9.8
mimeparse==0.1.3
mobile-codes==0.1.1
mock==1.0b1
mozilla-logger==0.2
newrelic==2.4.0.4
nose==1.2.1
oauth2==1.5.211
py-bcrypt==0.4
python-dateutil==1.5
python-memcached==1.48
raven==4.2.3
requests==0.14.1
schematic==0.2
six==1.2.0
slumber==0.5.3
statsd==1.0.0
suds==0.4
tower==0.3.4

# not on pypi
git+https://github.com/mozilla/nuggets.git@96e80a64aa4bfcfef4f43fc3ab6966450ccd7325#egg=nuggets
